extract graph component
