+++ airbyte-webapp/src/packages/cloud/views/credits/CreditsPage/components/UsagePerDayGraph.module.scss
@@ -1,0 +1,12 @@
+@use "scss/variables";
+
+.empty {
+  height: 100%;
+  padding-bottom: variables.$spacing-xl;
+}
+
+.container {
+  width: 100%;
+  height: 260px;
+  padding: 0 variables.$spacing-2xl variables.$spacing-xl 0;
+}
